third_party/WebKit/LayoutTests/fast/css/getComputedStyle/counterIncrement-without-counter-expected.txt

PASS style.counterIncrement is '';
PASS style.counterReset is '';
PASS successfullyParsed is true

TEST COMPLETE
